Refactor service Program.cs and update docs+scripts
