+++ service/Service/run.cmd
@@ -1,5 +1,5 @@
 @echo off
 
-dotnet restore
-dotnet build
-cmd /C "set ASPNETCORE_ENVIRONMENT=Development && dotnet run"
+dotnet clean
+dotnet build -c Debug -p "SolutionName=KernelMemory"
+cmd /C "set ASPNETCORE_ENVIRONMENT=Development && dotnet run --no-build --no-restore"
