+++ tools/README.md
@@ -1,4 +1,6 @@
-# upload-file.sh
+# Kernel memory web service scripts
+
+### upload-file.sh
 
 Simple client for command line uploads to Kernel Memory.
 
@@ -8,7 +10,7 @@ Instructions:
 ./upload-file.sh -h
 ```
 
-# ask.sh
+### ask.sh
 
 Simple client for asking questions about your documents from the command line.
 
@@ -18,7 +20,7 @@ Instructions:
 ./ask.sh -h
 ```
 
-# search.sh
+### search.sh
 
 Simple client for searching your indexed documents from the command line.
 
@@ -28,14 +30,41 @@ Instructions:
 ./search.sh -h
 ```
 
-# run-qdrant.sh
+# Vector DB scripts
+
+### run-elasticsearch.sh
+
+Script to start Elasticsearch using Docker for local development/debugging.
+
+Elasticsearch is used to store and search vectors, as an alternative to
+[Azure AI Search](https://azure.microsoft.com/products/ai-services/ai-search/).
+
+### run-mssql.sh
+
+Script to start MS SQL using Docker for local development/debugging.
+
+MS SQL is used to store and search vectors, as an alternative to
+[Azure AI Search](https://azure.microsoft.com/products/ai-services/ai-search/).
+
+### run-qdrant.sh
 
 Script to start Qdrant using Docker, for local development/debugging.
 
 Qdrant is used to store and search vectors, as an alternative to
 [Azure AI Search](https://azure.microsoft.com/products/ai-services/ai-search/).
 
-# run-rabbitmq.sh
+### run-redis.sh
+
+Script to start Redis using Docker, for local development/debugging.
+This will run Redis on port 6379, as well as running a popular Redis
+GUI, [RedisInsight](https://redis.com/redis-enterprise/redis-insight/), on port 8001.
+
+Redis is used to store and search vectors, as an alternative to
+[Azure AI Search](https://azure.microsoft.com/products/ai-services/ai-search/).
+
+# Orchestration queues scripts
+
+### run-rabbitmq.sh
 
 Script to start RabbitMQ using Docker, for local development/debugging.
 
@@ -43,10 +72,16 @@ RabbitMQ is used to provides queues for the asynchronous pipelines,
 as an alternative to
 [Azure Queues](https://learn.microsoft.com/azure/storage/queues/storage-queues-introduction).
 
-# run-redis.sh
+# Kernel memory runtime scripts
 
-Script to start Redis using Docker, for local development/debugging.
-This will run Redis on port 6379, as well as running a popular Redis GUI, [RedisInsight](https://redis.com/redis-enterprise/redis-insight/), on port 8001.
+### run-km-service.sh
 
-Redis is used to store and search vectors, as an alternative to
-[Azure AI Search](https://azure.microsoft.com/products/ai-services/ai-search/).
+Script to start KM service from source code, using KM nuget packages where configured as such.
+
+### run-km-service-from-source.sh
+
+Script to start KM service from local source code, ignoring KM nuget packages.
+
+### setup-km-service.sh
+
+Script to start KM service configuration wizard and create an appsettings.Development.json file.
