+++ tools/run-km-service.sh
@@ -3,6 +3,7 @@
 set -e
 
 cd "$(dirname "${BASH_SOURCE[0]:-$0}")"
+cd ../service/Service
 
 dotnet clean
 dotnet build -c Debug -p "SolutionName=KernelMemory"
