--- tools/run-service.sh
@@ -1,10 +1,0 @@
-#!/usr/bin/env bash
-
-set -e
-
-cd "$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)/"
-cd ../service/Service
-
-dotnet restore
-dotnet build
-ASPNETCORE_ENVIRONMENT=Development dotnet run
